+++ src/grid/cg_leaves.F90
@@ -137,7 +137,7 @@ contains
       implicit none
 
       class(cg_leaves_T),        intent(in) :: this       !< the list on which to perform the boundary exchange
-      integer,                   intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
+      integer(kind=4),           intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
       integer(kind=4), optional, intent(in) :: area_type  !< defines how do we treat boundaries
       integer(kind=4), optional, intent(in) :: bnd_type   !< Override default boundary type on external boundaries (useful in multigrid solver).
                                                           !< Note that BND_PER, BND_MPI, BND_SHE and BND_COR aren't external and cannot be overridden
@@ -164,7 +164,7 @@ contains
       implicit none
 
       class(cg_leaves_T),        intent(in) :: this       !< the list on which to perform the boundary exchange
-!      integer,                   intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
+!      integer(kind=4),           intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
 !      integer(kind=4), optional, intent(in) :: nb         !< number of grid cells to exchange
 !      integer(kind=4), optional, intent(in) :: area_type  !< defines how do we treat boundaries
 !      integer(kind=4), optional, intent(in) :: bnd_type   !< Override default boundary type on external boundaries (useful in multigrid solver).
@@ -195,7 +195,7 @@ contains
       implicit none
 
       class(cg_leaves_T),        intent(in) :: this       !< the list on which to perform the boundary exchange
-      integer,                   intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
+      integer(kind=4),           intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
       integer(kind=4), optional, intent(in) :: dim        !< do the internal boundaries only in the specified dimension
 
       type(cg_level_connected_T), pointer   :: curl
@@ -221,7 +221,7 @@ contains
       implicit none
 
       class(cg_leaves_T),        intent(in) :: this       !< the list on which to perform the boundary exchange
-      integer,                   intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
+      integer(kind=4),           intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
       integer(kind=4), optional, intent(in) :: dim        !< do the internal boundaries only in the specified dimension
 
       type(cg_level_connected_T), pointer   :: curl
@@ -247,7 +247,7 @@ contains
       implicit none
 
       class(cg_leaves_T),        intent(in) :: this       !< the list on which to perform the boundary exchange
-      integer,                   intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
+      integer(kind=4),           intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
       integer(kind=4), optional, intent(in) :: area_type  !< defines how do we treat boundaries
       integer(kind=4), optional, intent(in) :: bnd_type   !< Override default boundary type on external boundaries (useful in multigrid solver).
                                                           !< Note that BND_PER, BND_MPI, BND_SHE and BND_COR aren't external and cannot be overridden
@@ -274,7 +274,7 @@ contains
       implicit none
 
       class(cg_leaves_T),        intent(in) :: this       !< the list on which to perform the boundary exchange
-!      integer,                   intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
+!      integer(kind=4),           intent(in) :: ind        !< Negative value: index of cg%q(:) 3d array
 !      integer(kind=4), optional, intent(in) :: area_type  !< defines how do we treat boundaries
 !      integer(kind=4), optional, intent(in) :: bnd_type   !< Override default boundary type on external boundaries (useful in multigrid solver).
                                                           !< Note that BND_PER, BND_MPI, BND_SHE and BND_COR aren't external and cannot be overridden
